Extend waitForPage for admin app
